rename storage.Exceptions to storage.exceptions to better match PEP8
